--- site/src/component/NewReviewForm/ReviewForm.scss
@@ -1,71 +1,0 @@
-@use '../../globals';
-
-.review-form-modal {
-  .modal-dialog {
-    width: 346px;
-    display: flex;
-    flex-direction: column;
-    margin: auto;
-    overflow-y: auto;
-  }
-
-  .modal-header {
-    font-weight: bold;
-  }
-
-  .editing-notice {
-    margin-top: -16px;
-    font-size: 12px;
-    font-style: italic;
-  }
-
-  .form-check-input {
-    border: 1px solid rgba(136, 136, 136, 0.4);
-  }
-
-  .grade-difficulty-row {
-    display: flex;
-    gap: 16px;
-  }
-
-  .year-quarter-row {
-    display: flex;
-    gap: 16px;
-
-    > .form-group {
-      flex: 1;
-    }
-  }
-
-  .ppc-combobox {
-    .ppc-combobox__menu {
-      margin-top: 0;
-      border-radius: 0 0 8px 8px;
-      overflow: hidden;
-    }
-    .ppc-combobox__menu-list {
-      padding: 0;
-    }
-  }
-
-  .additional-details {
-    height: 90px;
-  }
-
-  .anonymous-checkbox {
-    margin-top: 16px;
-  }
-}
-
-.review-star-rating {
-  display: flex;
-  gap: 0.5rem;
-  cursor: pointer;
-}
-
-// monkey patch for recaptcha border bug
-.g-recaptcha {
-  overflow: hidden;
-  height: 76px;
-  border-radius: 4px;
-}
