+++ site/src/component/Review/Review.tsx
@@ -1,6 +1,6 @@
 import { FC, useState, useEffect, useCallback, useContext } from 'react';
 import SubReview from './SubReview';
-import ReviewForm from '../NewReviewForm/ReviewForm';
+import ReviewForm from '../ReviewForm/ReviewForm';
 // import ReviewForm from '../ReviewForm/ReviewForm';
 import './Review.scss';
